chore: refactor drawer header
