+++ packages/desktop/components/drawers/Drawer.svelte
@@ -1,10 +1,8 @@
 <script lang="ts">
     import { fade, fly } from 'svelte/transition'
-    import { Icon, Text, TextType } from '@ui'
-    import { localize } from '@core/i18n'
     import { NetworkConfigDrawerRouter } from '@components'
+    import DrawerHeader from './DrawerHeader.svelte'
     import { Router } from '@core/router'
-    import { Icon as IconEnum } from '@auxiliary/icon/enums'
     import { closeDrawer, DrawerDirection, DrawerId, drawerState } from '@desktop/auxilary/drawer'
 
     export let onClose: () => unknown = () => {}
@@ -51,12 +49,6 @@
             closeDrawer()
         }
     }
-
-    function onBackClick(): void {
-        if (drawerRouter) {
-            drawerRouter.previous()
-        }
-    }
 </script>
 
 {#if $drawerState.active}
@@ -73,30 +65,7 @@
             out:fly|local={{ ...direction, duration: DRAWER_ANIMATION_DURATION_MS }}
             class="bg-white dark:bg-gray-800 {position} {isVertical ? 'vertical' : 'horizontal'}"
         >
-            <div class="flex flex-row items-center mb-12">
-                {#if showBackButton}
-                    <button on:click={onBackClick} class="absolute top-6.5 focus:text-blue-500">
-                        <Icon
-                            icon={IconEnum.ArrowLeft}
-                            classes="text-gray-500 dark:text-white hover:text-gray-600 dark:hover:text-gray-100"
-                        />
-                    </button>
-                {/if}
-
-                <Text type={TextType.h4} classes="text-center {showBackButton ? 'ml-9' : ''}">
-                    {localize(`views.dashboard.drawers.${$drawerState?.id}.${drawerRoute}.title`)}
-                </Text>
-
-                {#if !$drawerState.hideClose}
-                    <button on:click={onCloseClick} class="absolute top-7 right-7 focus:text-blue-500">
-                        <Icon
-                            icon={IconEnum.Close}
-                            classes="text-gray-500 dark:text-white hover:text-gray-600 dark:hover:text-gray-100"
-                        />
-                    </button>
-                {/if}
-            </div>
-
+            <DrawerHeader {drawerRoute} {drawerRouter} onClose={onCloseClick} />
             {#if $drawerState.id === DrawerId.NetworkConfig}
                 <NetworkConfigDrawerRouter bind:drawerRoute bind:drawerRouter />
             {/if}
